[DEBUG] fix gcc warnings

+++ Sources/ARUPDATER_Manager.c
@@ -114,7 +114,7 @@ void ARUPDATER_Manager_Delete (ARUPDATER_Manager_t **managerPtrAddr)
     }
 }
 
-int ARUPDATER_Manager_PlfVersionIsUpToDate(ARUPDATER_Manager_t *manager, const char *const rootFolder, eARDISCOVERY_PRODUCT product, int version, int edition, int extension, const char *localVersionBuffer, int bufferSize, eARUPDATER_ERROR *error)
+int ARUPDATER_Manager_PlfVersionIsUpToDate(ARUPDATER_Manager_t *manager, const char *const rootFolder, eARDISCOVERY_PRODUCT product, int version, int edition, int extension, const char *localVersionBuffer, size_t bufferSize, eARUPDATER_ERROR *error)
 {
     eARUPDATER_ERROR err = ARUPDATER_OK;
